Thermal model averaging (#273)
This PR enables multimaterial simulations including temperature
influence.

TODO:

The test passes but we use a temp that is temp0 still
`HeatTransfer` and `ForceModelsMulti` need further adaptations/implementations for the thermal equation to work in multimaterial context. This will be done in a follow up
Tested with a bimaterial strip example

---------

Co-authored-by: Sam Reeve <6740307+streeve@users.noreply.github.com>